backend/requirements.txt
Flask
requests
beautifulsoup4
openai
python-dotenv
nltk
fastapi
uvicorn
python-multipart